utilizando Firebase para desplegar y validar productos

+++ src/Components/ItemDetailContainer/ItemDetailContainer.jsx
@@ -1,8 +1,9 @@
 import { useParams } from "react-router-dom";
 import { useEffect, useState } from "react";
 import ItemDetail from "../ItemDetail/ItemDetail";
-import { buscaId , fakeProduct} from "../../helpers/pideDatos";
+import { fakeProduct } from "../../helpers/pideDatos";
 import LeonParrillero from "../LeonParrillero/LeonParrillero";
+import { doc, getDoc, getFirestore } from "firebase/firestore";
 
 
 export default function ItemDetailContainer() {
@@ -10,33 +11,51 @@ export default function ItemDetailContainer() {
   const { itemId } = useParams();
   const [loading, setLoading] = useState(true);
 
+
   useEffect(() => {
-    buscaId(itemId)
-      .then((res) => setItem(res))
-      .catch((messageNF) => {
-        console.log(messageNF);
+    setLoading(true);
+
+    const db = getFirestore();
+
+    // Codigo para traer UN producto
+    const productoRef = doc(db, "Productos", itemId);
+
+    getDoc(productoRef).then((documento) => {
+      if (documento.exists()) {
+        setItem({ id: documento.id, ...documento.data() })
+
+        setLoading(false);
+
+      } else {
         setItem(fakeProduct);
-      })
-      .finally(() => setLoading(false));
-      
+        setLoading(false);
+      }
+    }).catch((error) => {
+      console.error("Error obteniendo el producto:", error);
+
+
+    });
+
+
   }, [itemId]);
+
+
+
 
   const greeting = "RR's Grill Store";
   const message = item
     ? "Detalles de tu producto:"
     : "El producto que buscas ya no está disponible";
-  
+
   return (
     <main>
-      
-      
 
       {loading ? (
         <>
           <LeonParrillero />
         </>
       ) : (
-        item && <ItemDetail producto={item} greeting={greeting} message= {message} />
+        item && <ItemDetail producto={item} greeting={greeting} message={message} />
       )}
